% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillvars.R
\name{fillvars}
\alias{fillvars}
\title{Fill election-specific variables in V-Dem data}
\usage{
fillvars(fill_na = TRUE)
}
\arguments{
\item{fill_na}{Whether the NA values for breaks in the electoral regime should be filled with zeros. The default = TRUE.}
}
\value{
A data frame with the most recent V-Dem dataset from the vdemdata package where election variables have 
been filled.
}
\description{
A function to carryforward election-specific variables in V-Dem provided no interruption in 
the electoral regime (v2x_elecreg).
}
\details{
Some variables in the V-Dem dataset are coded at the election-year only. However, users 
may wish to include these in country-year analyses. This function fills the election specific
indicators for subsequent years, provided that there has not been any interruption in the 
electoral regime, as defined by v2x_elecreg. 

More specifically, it carries forward values between elections for all variables that are 
election-specific, unless v2x_elecreg becomes zero. When v2x_elecreg becomes zero, 
users may either fill the resulting missing values with zeros (the smallest possible score).

Note: This function only fills variables with the suffix _osp, _ord, _mean, and _nr. The estimates
from the measurement model are not included because these do not have an absolute low point, on
their approximate z-score distributions. Future iterations of this function may provide additonal
options for filling these values with some theoretical minimum. 

Specifically, the variables filled include:
"v2elmulpar", "v2elpeace", "v2elrgstry", "v2elvotbuy", "v2elirreg", "v2elintim",  "v2elboycot", "v2elfrcamp", 
"v2elpdcamp", "v2elfrfair", "v2elaccept", "v2elasmoff".
}
