% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_overlap.R
\name{find_overlap}
\alias{find_overlap}
\title{Find the overlap between episodes of democratization and autocratization.}
\usage{
find_overlap(episodes = vdemdata::get_eps())
}
\arguments{
\item{episodes}{The outcome of get_eps(), episodes of regime transformation (ERT), to be used for finding potential overlaps (depending on individual parameter setting). By default with standard parameters.}
}
\value{
A data frame showing the country/year of overlaps between democratization and autocratization episodes.
}
\description{
Find the overlap between episodes of democratization and autocratization.
}
\examples{
#Don't run
#Find the overlap between democratization and autocratization episodes

#overlap <- find_overlap(episodes)

}
