% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_seqs_dem}
\alias{find_seqs_dem}
\title{Identify sequences of potential democratization episodes}
\usage{
find_seqs_dem(
  v,
  r,
  start_incl = 0,
  year_turn = 0,
  cum_turn = 0,
  tolerance = 0L
)
}
\description{
This is a subfunction (c++) of vdemdata::get_dem see the
documentation of vdemdata::get_dem (?get_dem) for details on
parameters, etc.
}
