% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_info.R
\name{find_var}
\alias{find_var}
\title{Search variable information by one or several keywords}
\usage{
find_var(keywords)
}
\arguments{
\item{keywords}{A character vector of length one with one or
several keywords separated by comma to search in
the name of the variable. You can also pass regex
notation (see \code{examples(find_var)}).}
}
\description{
\code{find_var} returns a data.frame with the Codebook
information on the indicators. The keywords are searched
in the names of variables.
}
\examples{
# Don't run
# df <- find_var("Democracy, ^elect")
# View(df)

}
