% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{fix_overlap}
\alias{fix_overlap}
\title{An interactive function to fix the overlap between episodes of democratization and autocratization.}
\usage{
fix_overlap(overlap, merged)
}
\arguments{
\item{overlap}{The outcome of the find_overlap() function. By default with standard parameters.}

\item{merged}{The outcome of get_aut(), autocratization episodes. By default with standard parameters.}
}
\value{
Recoding of overlapping episodes based on the interactive function.
}
\description{
An interactive function to fix the overlap between episodes of democratization and autocratization.
}
\examples{
Fix the overlap between democratization and autocratization episodes

fix_overlap(overlap, merged)

}
