% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aut.R
\name{get_aut}
\alias{get_aut}
\title{Get episodes of autocratization (democratic regression, autocratic regression)}
\usage{
get_aut(
  data = vdemdata::vdem,
  start_incl = -0.01,
  cum_incl = -0.1,
  year_turn = 0.02,
  cum_turn = 0.1,
  tolerance = 5,
  output = "df"
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set with filled election variables (fill_vdem)}

\item{start_incl}{A threshold for detecting the onset of "potential" autocratization episodes.
By default a negative change in the EDI (Vdem's Electoral Democracy Index) of at least -0.01 from year(t-1) to year(t).}

\item{cum_incl}{A threshold to identify a "manifest" episode of autocratization as a cumulative negative change of the EDI (Vdem's Electoral Democracy Index)
between the start and end of a sequence. By default a cumulative change of -0.1 on the EDI.}

\item{year_turn}{A threshold to identify a sudden "turn" during a year of an ongoing episode of autocratization (=failed autocratization).
By default a yearly change of +0.02 on the EDI (Vdem's Electoral Democracy Index).}

\item{cum_turn}{A threshold to identify a gradual "turn" during an ongoing episode of autocratization (=failed autocratization).
By default a cumulative change of +0.1 on the EDI (Vdem's Electoral Democcracy Index) between the start and end of a sequence.}

\item{tolerance}{A threshold to specify the number of "stasis" observations (\emph{i.e.}, observations neither increasing
or decreasing significantly) permitted before stopping a sequence. By default 5 years.}

\item{output}{The default output is a data frame. Alternatively, output = "summary" prints the overall observed numbers of each episode types to the console,
output = "table" compiles a short table showing each country with start and end year of each episode and episode ID.}
}
\value{
A data frame specifying episodes of autocratization in the most recent Vdem data set: democratic regression for those episodes starting in democracy ("aut_ep_dem") and
autocratic regression for those episodes starting in autocracy ("aut_ep_aut"), further distinguishing subtypes of democratic regression into (1) breakdown ("breakdown"), and (2) averted democratic regression ("averted").
}
\description{
Helps to identify episodes of autocratization in the most recent vdem data set.
\emph{Autocratization} is defined as any movement towards autocracy which starts within democracies or autocracies (cf. Lührmann and Lindberg, Democratization, 2019).
\emph{Democratic regression} is defined as a subtype of autocratization and specifically focuses on any movement towards autocracy
which starts in democracies. \emph{Autocratic regression} is also a subtype of autocratization and
concerns all those which are already autocratic and further decline (cf. Boese et al., forthcoming in Democratization, 2020).
}
\examples{
Get the episodes with standard parameters:
aut_episodes <- get_aut()

Get a short table with country, start and end year of episodes, episode ID only
aut_table <- get_aut(output="table")

Get the overall numbers of episode types printed to the console
aut_summary <- get_aut(output="summary")


}
