% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Get episodes of democratization (opening, democratic deepening)}
\usage{
get_dem(
  data = vdemdata::vdem,
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = -0.02,
  cum_turn = -0.1,
  tolerance = 10
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set.}

\item{start_incl}{A threshold for detecting the onset of "potential" democratization episodes.
By default a positive change in the EDI (Vdem's Electoral Democracy Index) of at least +0.01 from year(t-1) to year(t).}

\item{cum_incl}{A threshold to identify a "manifest" episode of democratization as a cumulative positive change of the EDI (Vdem's Electoral Democracy Index)
between the start and end of a sequence. By default a cumulative change of +0.1 on the EDI.}

\item{year_turn}{A threshold to identify a sudden "turn" during a year of an ongoing episode of democratization (=failed democratization).
By default a yearly change of -0.02 on the EDI (Vdem's Electoral Democracy Index).}

\item{cum_turn}{A threshold to identify a gradual "turn" during an ongoing episode of democratization (=failed democratization).
By default a cumulative change of -0.1 on the EDI (Vdem's Electoral Democcracy Index) between the start and end of a sequence.}

\item{tolerance}{A threshold to specify the number of "stasis" observations (\emph{i.e.}, observations neither increasing
or decreasing significantly) permitted before stopping a sequence. By default 10 years.}
}
\value{
A data frame specifying episodes of democratization in the most recent Vdem data set: democratic deepening for those episodes starting in democracy ("dem_ep_dem") and
opening for those episodes starting in autocracy ("dem_ep_aut"), further distinguishing successful opening episodes ("success"), and three types of failure,
(1) preempted ("fail_preem"), (2) reverted ("fail_rev"), and (3) stabilized autocracy ("fail_stab").
}
\description{
Helps to identify episodes of democratization (liberalization, democratic deepening) in the most recent vdem data set.
\emph{Democratization} is an umbrella term for any movement towards demcracy - be it in autocracies or democracies.
\emph{Opening} is defined as a subtype of democratiztion and specifically focuses on any movement towards democracy
which starts in autocracies. \emph{Democratic deepening} is also a subtype of democratization and
concerns all those which are already democratic and further improve their democratic traits (cf. Wilson et al., 2020).
}
\examples{
Get the episodes with standard parameters:
dem_episodes <- get_dem()

}
