% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eps.R
\name{get_eps}
\alias{get_eps}
\title{Get episodes of regime transformation (ERT)}
\usage{
get_eps(
  data = vdemdata::vdem,
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  cum_turn = 0.1,
  tolerance = 5
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set.}

\item{start_incl}{A threshold for detecting the onset of "potential" episodes.
By default a change in the EDI (Vdem's Electoral Democracy Index) of at least +/-0.01 from year(t-1) to year(t).}

\item{cum_incl}{A threshold to identify a "manifest" episodes as a cumulative change of the EDI (Vdem's Electoral Democracy Index)
between the start and end of a sequence. By default a cumulative change of +/-0.1 on the EDI.}

\item{year_turn}{A threshold to identify a sudden "turn" during a year of an ongoing episode (=failed democratization/autocratization).
By default a yearly change of +/-0.03 on the EDI (Vdem's Electoral Democracy Index). Note: Advanced users who wish to remove this criteria altogether
should set the value of year turn equal to cum turn. Setting this to zero would allow for an episode to terminate when any year of no change is encountered.}

\item{cum_turn}{A threshold to identify a gradual "turn" during an ongoing episode (=failed democratization/autocratization).
By default a cumulative change of -0.1 on the EDI (Vdem's Electoral Democcracy Index) between the start and end of a sequence.}

\item{tolerance}{A threshold to specify the number of "stasis" observations (\emph{i.e.}, observations neither increasing
or decreasing significantly) permitted before stopping a sequence. By default 5 years.}
}
\value{
A data frame specifying episodes of regime transformation in the most recent Vdem data set.

Democratization episodes: democratic deepening for those episodes starting in democracy ("dem_ep_dem") and
liberalization for those episodes starting in autocracy ("dem_ep_aut"), further distinguishing successful episodes of democratic transitions ("success"), and three types of failure,
(1) preempted ("fail_preem"), (2) reverted ("fail_rev"), and (3) stabilized autocracy ("fail_stab").

Autocratization episodes: democratic regression for those episodes starting in democracy ("aut_ep_dem") and
autocratic regression for those episodes starting in autocracy ("aut_ep_aut"), further distinguishing subtypes of democratic regression into (1) breakdown ("breakdown"), and (2) averted democratic regression ("averted").
}
\description{
Helps to identify episodes of democratization (liberalization, democratic deepening) and autocratization (demcratic regression, autocratic regression) in the most recent vdem data set.
}
\details{
\emph{Democratization} is an umbrella term for any movement towards demcracy - be it in autocracies or democracies.
\emph{liberalization} is defined as a subtype of democratiztion and specifically focuses on any movement towards democracy
which starts in autocracies. \emph{Democratic deepening} is also a subtype of democratization and
concerns all those which are already democratic and further improve their democratic traits (cf. Wilson et al., 2020).

\emph{Autocratization} is defined as any movement towards autocracy which starts within democracies or autocracies (cf. Lührmann and Lindberg, Democratization, 2019).
\emph{Democratic regression} is defined as a subtype of autocratization and specifically focuses on any movement towards autocracy
which starts in democracies. \emph{Autocratic regression} is also a subtype of autocratization and
concerns all those which are already autocratic and further decline (cf. Boese et al., forthcoming in Democratization, 2020).
}
\examples{
#Don't run
#Get the episodes with standard parameters:
#episodes <- get_eps()

}
