% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_episodes.R
\name{plot_episodes}
\alias{plot_episodes}
\title{Plot Episodes of Regime Transformation (ERT) over time.}
\usage{
plot_episodes(
  abs = T,
  years = c(1900, 2020),
  country = c(),
  start_incl = 0.01,
  cum_incl = 0.1,
  year_turn = 0.03,
  cum_turn = 0.1,
  tolerance = 5
)
}
\arguments{
\item{abs}{Logical value: if TRUE, the absolute number of countries in an episode for each year is plotted. If FALSE, the share
of countries (in \%) undergoing democratization or autocratization is plotted.}

\item{years}{Vector with numeric values indicating the minimum and maximum year to be plotted.}

\item{country}{Character vector containing the country for which episodes should be shown. Only countries from the
country_name column in the V-Dem data set are accepted. If no countries are specified, global episode numbers
are plotted. Only a single country can be selected.}

\item{start_incl}{A threshold for detecting the onset of "potential" episodes.
By default a change in the EDI (Vdem's Electoral Democracy Index) of at least +/-0.01 from year(t-1) to year(t).}

\item{cum_incl}{A threshold to identify a "manifest" episodes as a cumulative change of the EDI (Vdem's Electoral Democracy Index)
between the start and end of a sequence. By default a cumulative change of +/-0.1 on the EDI.}

\item{year_turn}{A threshold to identify a sudden "turn" during a year of an ongoing episode (=failed democratization/autocratization).
By default a yearly change of +/-0.03 on the EDI (Vdem's Electoral Democracy Index). Note: Advanced users who wish to remove this criteria altogether
should set the value of year turn equal to cum turn. Setting this to zero would allow for an episode to terminate when any year of no change is encountered.}

\item{cum_turn}{A threshold to identify a gradual "turn" during an ongoing episode (=failed democratization/autocratization).
By default a cumulative change of -0.1 on the EDI (Vdem's Electoral Democcracy Index) between the start and end of a sequence.}

\item{tolerance}{A threshold to specify the number of "stasis" observations (\emph{i.e.}, observations neither increasing
or decreasing significantly) permitted before stopping a sequence. By default 5 years.}
}
\value{
The output of this function is a [ggplot2:ggplot()] object with global or country-specific episodes.
}
\description{
`plot_episodes` plots Episodes of Regime Transformation (ERT) over time. It shows either
the global number/share of countries undergoing democratization and autocratization or
episodes for an individual country.
}
\details{
This function is a wrapper for [ggplot2:ggplot()] and produces a plot that shows
democratization and autocratization episodes over time.
Either the global number/share of countries undergoing democratization and autocratization or
episodes for an individual country are plotted. The function calls the [vdemdata:get_eps()] funtion that is
used to identify episodes.
}
\examples{
\dontrun{
# Plot autocratization and democratization episodes over time using
 default values from [vdemdata:get_eps()].

 plot_episodes()

# Plot episodes for Belgium between 1910 and 2010.

 plot_episodes(country = c("Belgium"),
               years = c(1910, 2010))
}
}
