% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_info.R
\name{var_info}
\alias{var_info}
\title{Shows basic information on a specific variable as given in the codebook}
\usage{
var_info(var_tag)
}
\arguments{
\item{var_tag}{A variable's name as used in the V-Dem data set,
e.g. "v2x_polyarchy" for the electoral democracy index.
Note: use quotation marks vor "var_tag".}
}
\value{
A list of information on the desired variable.
}
\description{
Provides information on full name as specified in the codebook, variable type,
data collection process (questions, clarifications, responses),
scale, aggregation, years and sources (plus notes)
of a variable in the V-Dem data set
}
\examples{
# Get information on the variable "v2x_polyarchy" (the electoral democracy index)
var_info("v2x_polyarchy")
}
