% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{V-Dem data}
\format{
A data frame with 27013 observations and 4108 variables.
}
\source{
https://www.v-dem.net/
}
\usage{
vdem
}
\description{
Loads the most recent version of the V-Dem data set.
}
\details{
Varieties of Democracy (V-Dem) is a new approach to conceptualizing
and measuring democracy. We provide a multidimensional and disaggregated
dataset that reflects the complexity of the concept of democracy as a
system of rule that goes beyond the simple presence of elections.
The V-Dem project distinguishes between five high-level principles of
democracy: electoral, liberal, participatory, deliberative, and
egalitarian, and collects data to measure these principles.
With six Principal Investigators (PIs), seventeen Project Managers
(PMs) with special responsibility for issue areas,
more than thirty Regional Managers (RMs),
170 Country Coordinators (CCs), Research Assistants,
and 3,000 Country Experts (CEs), the V-Dem project is one
of the largest social science data collection projects
focusing on research. The Headquarters is based at the
V-Dem Institute, the Department of Political Science at the
University of Gothenburg, Sweden. The most recent data set (v9)
covers 202 countries from 1789-2018 and includes 450+ indicators,
81 indices and 5 high-level indices.
For more info see https://www.v-dem.net/
}
\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Anna Luhrmann, Kyle L. Marquardt, Kelly McMann, Pamela Paxton, Daniel Pemstein, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Steven Wilson, Agnes Cornell, Nazifa Alizada, Lisa Gastaldi, Haakon Gjerløw, Garry Hindle, Nina Ilchenko, Laura Maxwell, Valeriya Mechkova, Juraj Medzihorsky, Johannes von Römer, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig, and Daniel Ziblatt. 2020. ”V-Dem [Country–Year/Country–Date] Dataset v10”. Varieties of Democracy (V-Dem) Project. https://doi.org/10.23696/vdemds20.
}
\keyword{datasets}
