% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdemdata.R
\docType{package}
\name{vdemdata}
\alias{vdemdata}
\title{vdemdata}
\description{
A package to load, expore and work with the most recent V-Dem (Varieties of Democracy) dataset.
}
\references{
\url{https://github.com/vdeminstitute/vdemdata} and \url{https://www.v-dem.net/en/}
}
\author{
Seraphine F. Maerz (package creater and maintainer, seraphine.maerz@v-dem.net), Amanda B. Edgell, Laura Maxwell, Sebastian Hellmeier, Nina Ilchenko
}
