% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vparty}
\alias{vparty}
\title{V-Party data}
\format{
A data frame with 11898 observations and 384 variables.
}
\source{
https://www.v-dem.net/
}
\usage{
vparty
}
\description{
Loads the most recent version of the V-Party dataset.
}
\details{
Varieties of Democracy (V-Dem) is a new approach to conceptualizing
and measuring democracy. We provide a multidimensional and disaggregated
dataset that reflects the complexity of the concept of democracy as a
system of rule that goes beyond the simple presence of elections.
The V-Dem project distinguishes between five high-level principles of
democracy: electoral, liberal, participatory, deliberative, and
egalitarian, and collects data to measure these principles.
With five Principal Investigators (PIs), thirty-nine Project Managers
(PMs) with special responsibility for issue areas,
more than thirty Regional Managers (RMs),
134 Country Coordinators (CCs), Research Assistants,
and over 4,000 Country Experts (CEs). The V-Dem project is one
of the largest social science data collection projects
focusing on research. The Headquarters is based at the
V-Dem Institute, the Department of Political Science at the
University of Gothenburg, Sweden. The most recent data set (v15)
covers 202 countries from 1789-2024 and includes 500+ indicators,
81 indices and 5 high-level indices.
For more info see https://www.v-dem.net/
}
\references{
Lindberg, Staffan I., Nils Düpont, Masaaki Higashijima, Yaman Berker Kavasoglu, Kyle L. Marquardt, Michael Bernhard, Holger Döring, Allen Hicken, Melis Laebens, Juraj Medzihorsky, Anja Neundorf, Ora John Reuter, Saskia Ruth–Lovell, Keith R. Weghorst, Nina Wiesehomeier, Joseph Wright, Nazifa Alizada, Paul Bederke, Lisa Gastaldi, Sandra Grahn, Garry Hindle, Nina Ilchenko, Johannes von Römer, Steven Wilson, Daniel Pemstein, Brigitte Seim. 2022. Varieties of Party Identity and Organization (V–Party) Dataset V2. Varieties of Democracy (V–Dem) Project. https://www.v-dem.net/data/v-party-dataset/
}
\keyword{datasets}
