% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps.R
\name{find_overlap}
\alias{find_overlap}
\title{Find the overlap between episodes of democratization and autocratization.}
\usage{
find_overlap(
  dem_ep = vdemdata::get_dem(),
  aut_ep = vdemdata::get_aut(),
  merged = dem_ep \%>\% left_join(aut_ep)
)
}
\arguments{
\item{dem_ep}{The outcome of get_dem(), democratization episodes. By default with standard parameters.}

\item{aut_ep}{The outcome of get_aut(), autocratization episodes. By default with standard parameters.}
}
\value{
A data frame showing the country/year of overlaps between democratization and autocratization episodes.
}
\description{
Find the overlap between episodes of democratization and autocratization.
}
\examples{
Find the overlap between democratization and autocratization episodes

democratization_episodes <- get_dem()
autocratization_episodes <- get_aut()

overlap <- find_overlap(democratization_episodes, autocratization_episodes)

}
