% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_seqs_aut}
\alias{find_seqs_aut}
\title{Identify sequences of potential autocratization episodes}
\usage{
find_seqs_aut(
  v,
  r,
  start_incl = 0,
  year_turn = 0,
  cum_turn = 0,
  tolerance = 0L
)
}
\description{
This is a subfunction (c++) of vdemdata::get_aut
see the documentation of vdemdata::get_aut (?get_aut) for details on parameters, etc.
}
