% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_seqs_dem}
\alias{find_seqs_dem}
\title{Identify sequences of potential democratization episodes}
\usage{
find_seqs_dem(
  v,
  r,
  t,
  start_incl = 0.01,
  year_turn = -0.03,
  cum_turn = -0.1,
  tolerance = 5L
)
}
\description{
This is a subfunction (c++) of vdemdata::get_dem see the
documentation of vdemdata::get_dem (?get_dem) for details on
parameters, etc.
}
