% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_overlap.R
\name{fix_overlap}
\alias{fix_overlap}
\title{An interactive function to fix the overlap between episodes of democratization and autocratization.}
\usage{
fix_overlap(
  episodes = vdemdata::get_eps(),
  overlap = vdemdata::find_overlap(episodes)
)
}
\arguments{
\item{episodes}{The outcome of get_eps(), episodes of regime transformation (ERT), to be used for fixing potential overlaps (depending on individual parameter setting). By default with standard parameters.}

\item{overlap}{The outcome of find_overlap(merged), depending on individual parameter settings.}
}
\value{
A "merged" dataframe after recoding overlapping episodes based on the interactive function.
}
\description{
An interactive function to fix the overlap between episodes of democratization and autocratization.
}
\examples{
#Don't run
#Fix the overlap between democratization and autocratization episodes

#fix_overlap(episodes)

}
