% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_overlap.R
\name{fix_overlap}
\alias{fix_overlap}
\title{An interactive function to fix the overlap between episodes of democratization and autocratization.}
\usage{
fix_overlap(dem_ep = vdemdata::get_dem(), aut_ep = vdemdata::get_aut())
}
\arguments{
\item{dem_ep}{The outcome of get_dem(), democratization episodes. By default with standard parameters.}

\item{aut_ep}{The outcome of get_aut(), autocratization episodes. By default with standard parameters.}
}
\value{
Recoding of overlapping episodes based on the interactive function.
}
\description{
An interactive function to fix the overlap between episodes of democratization and autocratization.
}
\examples{
#Don't run
#Fix the overlap between democratization and autocratization episodes

#fix_overlap(dem_ep, aut_ep)

}
