% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aut.R
\name{get_aut}
\alias{get_aut}
\title{Get episodes of autocratization (democratic backsliding, closing)}
\usage{
get_aut(
  data = vdemdata::fill_vdem,
  start_incl = -0.01,
  cum_incl = -0.1,
  year_turn = 0.02,
  cum_turn = 0.1,
  tolerance = 10
)
}
\arguments{
\item{data}{The data based on which the episodes are identified.
By default the most recent vdem data set with filled election variables (fill_vdem)}

\item{start_incl}{A threshold for detecting the onset of "potential" autocratization episodes.
By default a negative change in the EDI (Vdem's Electoral Democracy Index) of at least -0.01 from year(t-1) to year(t).}

\item{cum_incl}{A threshold to identify a "manifest" episode of autocratization as a cumulative negative change of the EDI (Vdem's Electoral Democracy Index)
between the start and end of a sequence. By default a cumulative change of -0.1 on the EDI.}

\item{year_turn}{A threshold to identify a sudden "turn" during a year of an ongoing episode of autocratization (=failed autocratization).
By default a yearly change of +0.02 on the EDI (Vdem's Electoral Democracy Index).}

\item{cum_turn}{A threshold to identify a gradual "turn" during an ongoing episode of autocratization (=failed autocratization).
By default a cumulative change of +0.1 on the EDI (Vdem's Electoral Democcracy Index) between the start and end of a sequence.}

\item{tolerance}{A threshold to specify the number of "stasis" observations (\emph{i.e.}, observations neither increasing
or decreasing significantly) permitted before stopping a sequence. By default 10 years.}
}
\value{
A data frame specifying episodes of autocratization in the most recent Vdem data set: democratic backsliding for those episodes starting in democracy ("aut_ep_dem") and
closing for those episodes starting in autocracy ("aut_ep_aut"), further distinguishing subtypes of democratic backsliding (1) breakdown ("breakdown"), and (2) averted democratic backsliding ("averted").
}
\description{
Helps to identify episodes of autocratization in the most recent vdem data set.
\emph{Autocratization} is defined as any movement towards autocracy which starts within democracies or autocracies (cf. Lührmann and Lindberg, Democratization, 2019).
\emph{Democratic backsliding} is defined as a subtype of autocratization and specifically focuses on any movement towards autocracy
which starts in democracies. \emph{Closing} is also a subtype of autocratization and
concerns all those which are already autocratic and further decline (cf. Boese et al., forthcoming in Democratization, 2020).
}
\examples{
Get the episodes with standard parameters:
aut_episodes <- get_aut()

}
