% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdemdata.R
\docType{package}
\name{vdemdata}
\alias{vdemdata}
\title{vdemdata}
\description{
A package to load, expore and work with the most recent V-Dem (Varieties of Democracy) dataset.
}
\author{
Seraphine Maerz <seraphine.maerz@gu.se>
}
